Refactor/renaming classes

+++ InstaConnect.Business/Services/FollowService.cs
@@ -15,18 +15,21 @@ namespace InstaConnect.Business.Services
         private readonly IMapper _mapper;
         private readonly IResultFactory _resultFactory;
         private readonly IFollowRepository _followRepository;
-        private readonly IInstaConnectUserManager _instaConnectUserManager;
+        private readonly IUserRepository _userRepository;
+        private readonly IAccountManager _accountManager;
 
         public FollowService(
             IMapper mapper,
             IResultFactory resultFactory,
             IFollowRepository followRepository,
-            IInstaConnectUserManager instaConnectUserManager)
+            IUserRepository userRepository,
+            IAccountManager accountManager)
         {
             _mapper = mapper;
             _resultFactory = resultFactory;
-            _followRepository = followRepository;
-            _instaConnectUserManager = instaConnectUserManager;
+            _followRepository = followRepository; ;
+            _userRepository = userRepository;
+            _accountManager = accountManager;
         }
 
         public async Task<IResult<ICollection<FollowResultDTO>>> GetAllAsync(string followerId, string followingId, int page, int amount)
@@ -79,9 +82,18 @@ namespace InstaConnect.Business.Services
             return okResult;
         }
 
-        public async Task<IResult<FollowResultDTO>> AddAsync(FollowAddDTO followAddDTO)
+        public async Task<IResult<FollowResultDTO>> AddAsync(string followerId, FollowAddDTO followAddDTO)
         {
-            var existingFollower = await _instaConnectUserManager.FindByIdAsync(followAddDTO.FollowerId);
+            var validUser = _accountManager.ValidateUser(followerId, followAddDTO.FollowerId);
+
+            if (!validUser)
+            {
+                var forbiddenResult = _resultFactory.GetForbiddenResult<FollowResultDTO>();
+
+                return forbiddenResult;
+            }
+
+            var existingFollower = await _userRepository.FindEntityAsync(f => f.Id == followAddDTO.FollowerId);
 
             if (existingFollower == null)
             {
@@ -90,7 +102,7 @@ namespace InstaConnect.Business.Services
                 return badRequestResult;
             }
 
-            var existingFollowing = await _instaConnectUserManager.FindByIdAsync(followAddDTO.FollowingId);
+            var existingFollowing = await _userRepository.FindEntityAsync(f => f.Id == followAddDTO.FollowingId);
 
             if (existingFollowing == null)
             {
@@ -136,13 +148,22 @@ namespace InstaConnect.Business.Services
 
         public async Task<IResult<FollowResultDTO>> DeleteAsync(string followerId, string id)
         {
-            var existingFollow = await _followRepository.FindEntityAsync(f => f.Id == id && f.FollowerId == followerId);
+            var existingFollow = await _followRepository.FindEntityAsync(f => f.Id == id);
 
             if (existingFollow == null)
             {
                 var notFoundResult = _resultFactory.GetNotFoundResult<FollowResultDTO>();
 
                 return notFoundResult;
+            }
+
+            var validUser = _accountManager.ValidateUser(followerId, existingFollow.FollowerId);
+
+            if (!validUser)
+            {
+                var forbiddenResult = _resultFactory.GetForbiddenResult<FollowResultDTO>();
+
+                return forbiddenResult;
             }
 
             await _followRepository.DeleteAsync(existingFollow);
