+++ InstaConnect.Business/Services/MessageService.cs
@@ -2,6 +2,7 @@
 using InstaConnect.Business.Abstraction.Factories;
 using InstaConnect.Business.Abstraction.Helpers;
 using InstaConnect.Business.Abstraction.Services;
+using InstaConnect.Business.Models.DTOs.Follow;
 using InstaConnect.Business.Models.DTOs.Message;
 using InstaConnect.Business.Models.Results;
 using InstaConnect.Business.Models.Utilities;
@@ -17,20 +18,23 @@ namespace InstaConnect.Business.Services
         private readonly IResultFactory _resultFactory;
         private readonly IMessageRepository _messageRepository;
         private readonly IMessageSender _messageSender;
-        private readonly IInstaConnectUserManager _instaConnectUserManager;
+        private readonly IUserRepository _userRepository;
+        private readonly IAccountManager _accountManager;
 
         public MessageService(
             IMapper mapper,
             IResultFactory resultFactory,
             IMessageRepository messageRepository,
             IMessageSender messageSender,
-            IInstaConnectUserManager instaConnectUserManager)
+            IUserRepository userRepository,
+            IAccountManager accountManager)
         {
             _mapper = mapper;
             _resultFactory = resultFactory;
             _messageRepository = messageRepository;
             _messageSender = messageSender;
-            _instaConnectUserManager = instaConnectUserManager;
+            _userRepository = userRepository;
+            _accountManager = accountManager;
         }
 
         public async Task<IResult<ICollection<MessageResultDTO>>> GetAllBySenderIdAndReceiverIdAsync(string senderId, string receiverId)
@@ -45,9 +49,9 @@ namespace InstaConnect.Business.Services
             return okResult;
         }
 
-        public async Task<IResult<MessageResultDTO>> GetByIdAsync(string userId, string id)
+        public async Task<IResult<MessageResultDTO>> GetByIdAsync(string senderId, string id)
         {
-            var existingMessage = await _messageRepository.FindEntityAsync(m => m.Id == id && m.SenderId == userId);
+            var existingMessage = await _messageRepository.FindEntityAsync(m => m.Id == id);
 
             if (existingMessage == null)
             {
@@ -56,15 +60,33 @@ namespace InstaConnect.Business.Services
                 return notFoundResult;
             }
 
+            var validUser = _accountManager.ValidateUser(senderId, existingMessage.SenderId);
+
+            if (!validUser)
+            {
+                var forbiddenResult = _resultFactory.GetForbiddenResult<MessageResultDTO>();
+
+                return forbiddenResult;
+            }
+
             var messageResultDTO = _mapper.Map<MessageResultDTO>(existingMessage);
             var okResult = _resultFactory.GetOkResult(messageResultDTO);
 
             return okResult;
         }
 
-        public async Task<IResult<MessageResultDTO>> AddAsync(MessageAddDTO messageAddDTO)
+        public async Task<IResult<MessageResultDTO>> AddAsync(string senderId, MessageAddDTO messageAddDTO)
         {
-            var existingSender = await _instaConnectUserManager.FindByIdAsync(messageAddDTO.SenderId);
+            var validUser = _accountManager.ValidateUser(senderId, messageAddDTO.SenderId);
+
+            if (!validUser)
+            {
+                var forbiddenResult = _resultFactory.GetForbiddenResult<MessageResultDTO>();
+
+                return forbiddenResult;
+            }
+
+            var existingSender = await _userRepository.FindEntityAsync(f => f.Id == messageAddDTO.SenderId);
 
             if (existingSender == null)
             {
@@ -73,7 +95,7 @@ namespace InstaConnect.Business.Services
                 return badRequestResult;
             }
 
-            var existingRecipient = await _instaConnectUserManager.FindByIdAsync(messageAddDTO.ReceiverId);
+            var existingRecipient = await _userRepository.FindEntityAsync(f => f.Id == messageAddDTO.ReceiverId);
 
             if (existingRecipient == null)
             {
@@ -92,15 +114,24 @@ namespace InstaConnect.Business.Services
             return noContentResult;
         }
 
-        public async Task<IResult<MessageResultDTO>> UpdateAsync(string userId, string id, MessageUpdateDTO messageUpdateDTO)
+        public async Task<IResult<MessageResultDTO>> UpdateAsync(string senderId, string id, MessageUpdateDTO messageUpdateDTO)
         {
-            var existingMessage = await _messageRepository.FindEntityAsync(m => m.Id == id && m.SenderId == userId);
+            var existingMessage = await _messageRepository.FindEntityAsync(m => m.Id == id);
 
             if (existingMessage == null)
             {
                 var notFoundResult = _resultFactory.GetNotFoundResult<MessageResultDTO>();
 
                 return notFoundResult;
+            }
+
+            var validUser = _accountManager.ValidateUser(senderId, existingMessage.SenderId);
+
+            if (!validUser)
+            {
+                var forbiddenResult = _resultFactory.GetForbiddenResult<MessageResultDTO>();
+
+                return forbiddenResult;
             }
 
             _mapper.Map(messageUpdateDTO, existingMessage);
@@ -111,15 +142,24 @@ namespace InstaConnect.Business.Services
             return noContentResult;
         }
 
-        public async Task<IResult<MessageResultDTO>> DeleteAsync(string userId, string id)
+        public async Task<IResult<MessageResultDTO>> DeleteAsync(string senderId, string id)
         {
-            var existingMessage = await _messageRepository.FindEntityAsync(m => m.Id == id && m.SenderId == userId);
+            var existingMessage = await _messageRepository.FindEntityAsync(m => m.Id == id);
 
             if (existingMessage == null)
             {
                 var notFoundResult = _resultFactory.GetNotFoundResult<MessageResultDTO>();
 
                 return notFoundResult;
+            }
+
+            var validUser = _accountManager.ValidateUser(senderId, existingMessage.SenderId);
+
+            if (!validUser)
+            {
+                var forbiddenResult = _resultFactory.GetForbiddenResult<MessageResultDTO>();
+
+                return forbiddenResult;
             }
 
             await _messageRepository.DeleteAsync(existingMessage);
